Add version override through environment variable
The environment variable `GH_CODEQL_VERSION` can be used to override the
global and local version specification. This is needed when we want to
override the CodeQL version used by the VS Code extension, because the
working directory used by the extension is `/` so we can't find any
local version specification part of the project using CodeQL.

With this override we can alias `code` to check for `.codeql-version` in
PWD or the path passed as an argument and set `GH_CODEQL_VERSION`.