Add functionality to enable use with the VS Code CodeQL extension (#20)
* Add local version pinning

This enables pinning a CodeQL version per working directory.

* Add install stub command

This installs a stub allowing the GH CodeQL extension to be used by the
VS Code extension.

* Add version override through environment variable

The environment variable `GH_CODEQL_VERSION` can be used to override the
global and local version specification. This is needed when we want to
override the CodeQL version used by the VS Code extension, because the
working directory used by the extension is `/` so we can't find any
local version specification part of the project using CodeQL.

With this override we can alias `code` to check for `.codeql-version` in
PWD or the path passed as an argument and set `GH_CODEQL_VERSION`.

* Update README with new commands

* Add description of local version pinning

* Add description of the `install-stub` command.

* Address incorrect indentation

* Add test for `install-stub` command

* Ensure the latest version is avaible for the test

* Add test for version override

* Restore the channel after the nightly version test

* Address spelling mistake in comment

* Remove superfluous download step

The selected version will be downloaded if not available.

* Address spelling mistake in comment

* Update install-stub command's descriptive comment

Explain what the command does.

* Address `jq` parser error

The version override causes a CodeQL CLI download that outputs cURL
outpuut which confuses the `jq` parser.
By running the command twice we remediate the issue.

* Add test for local pinned version modification

This test validates if we properly handle a local pinned version
modification without using the `set-local-version` command.

* Address pinning persistence influencing other tests

After each test using local pinning we unpin the version.

* Update error message when installing stub

When a directory doesn't exists, the user is asked to provide a
different directory to install a stub.

* Add permission check to install-stub command

Check if we can write to the provided/default directory and return an error message if we
can't.

* Disable per directory pinning by default

The command 'set-local-version' will show a warning when per directory
pinning is disabled (the default) and explicitly asks the user to enable
it.

* Update how we support local version

- Add explicit commands to enable/disable local version support.
- Show a warning when a local version is specified, but local version
  support is disabled.
- Show a warning when enabling local version support.
- Show an error when a local version is set while local version support
  is disabled.

* Address spelling mistake in test name

* Improve error message 'local-version' command